DRY with DiorySwiper & ContentSlide
